MISC: added 2 additional methods to User class - future_budget() and current_month_spending()
