+++ project/app/user.py
@@ -1,6 +1,7 @@
 import plotly.express as px
 import plotly.graph_objects as go
 import pandas as pd
+import numpy as np
 from datetime import timedelta
 
 
@@ -40,6 +41,89 @@ def get_last_time_period(transaction_df, time_period='week'):
 
     return subset
 
+def weighted_avg(data):
+    """
+    Given a dataframe
+    Return the datafram with a new column of weighted averages
+    """
+    categories = data.index
+    # get number of timepoints/observations and create weights
+    N = data.shape[1]
+    weights = [i for i in range(N ,0,-1)]
+    averages = []
+    # for each categorie, calculate the weighted average and 
+    # store in the averages list
+    for cat in categories:
+        cat_data = list(data.loc[cat])
+        # replace nan vaalues with 0 
+        for i in range(len(cat_data)):
+            if str(cat_data[i]) == 'nan':
+                cat_data[i] = 0
+        avg = np.average(cat_data, weights=weights)
+        averages.append(avg)
+
+    data['mean'] = averages
+    data = data.round()
+    return data
+
+def monthly_avg_spending(user_expenses_df, num_months=6, category='grandparent_category_name', weighted=True):
+    ticker = 0
+    cur_month = user_expenses_df['date'].max().month
+    cur_year = user_expenses_df['date'].max().year
+
+    while ticker < num_months:
+        # on first iteration
+        if ticker == 0:
+
+            if cur_month == 1:
+                prev_month = 12
+                user_exp_prev = user_expenses_df[(user_expenses_df['date'].dt.month == (
+                    prev_month)) & (user_expenses_df['date'].dt.year == (cur_year - 1))]
+                prev = user_exp_prev.groupby([category]).sum()
+                cur_month = prev_month
+                cur_year -= 1
+
+            else:
+                prev_month = cur_month - 1
+                user_exp_prev = user_expenses_df[(user_expenses_df['date'].dt.month == (
+                    prev_month)) & (user_expenses_df['date'].dt.year == cur_year)]
+                prev = user_exp_prev.groupby([category]).sum()
+                cur_month -= 1
+
+            datestring = f"{prev_month}/{str(cur_year)[2:]}"
+            prev.rename(columns={'amount_dollars': datestring}, inplace=True)
+            ticker += 1
+
+        else:
+            if cur_month == 1:
+                prev_month = 12
+                other = user_expenses_df[(user_expenses_df['date'].dt.month == (
+                    prev_month)) & (user_expenses_df['date'].dt.year == (cur_year - 1))]
+                other = other.groupby([category]).sum()
+                prev = pd.concat([prev, other], axis=1, sort=True)
+                cur_month = prev_month
+                cur_year -= 1
+            else:
+                prev_month = cur_month - 1
+                user_exp_prev = user_expenses_df[(user_expenses_df['date'].dt.month == (
+                    prev_month)) & (user_expenses_df['date'].dt.year == cur_year)]
+                other = user_exp_prev.groupby([category]).sum()
+                prev = pd.concat([prev, other], axis=1, sort=True)
+                cur_month -= 1
+
+            datestring = f"{prev_month}/{str(cur_year)[2:]}"
+            prev.rename(columns={'amount_dollars': datestring}, inplace=True)
+            ticker += 1
+
+    prev = prev.drop(columns=['amount_cents'])
+    
+    if weighted:
+        prev = weighted_avg(prev)
+    else:
+        prev['mean'] = round(prev.mean(axis=1))
+
+    return prev
+
 
 class User():
     def __init__(self, id, transactions, name=None, show=False):
@@ -61,6 +145,7 @@ class User():
         self.expenses = self.data[(self.data['grandparent_category_name'] != 'Transfers') & (
             self.data['amount_cents'] > 0)]
         self.show = show
+        self.past_months = 6
 
     def get_user_data(self):
         """
@@ -199,3 +284,64 @@ class User():
             fig.show()
 
         return fig.to_json()
+    
+    def future_budget(self, monthly_savings_goal=50, num_months=6, weighted=True):
+
+        # get dataframe of average spending per category over last 6 months
+        avg_spending_by_month_df = monthly_avg_spending(
+            self.expenses, num_months=num_months, weighted=weighted)
+        
+        # turn into dictionary where key is category and value is average spending
+        # . for that category
+        avg_cat_spending_dict = dict(avg_spending_by_month_df['mean'])
+
+        # label disctionary columns
+        discretionary = ['Food', 'Recreation', 'Shopping', 'Other']
+
+        # add column to df where its True if category is discretionary and False
+        # . otherwise
+        avg_spending_by_month_df['disc'] = [
+            True if x in discretionary else False for x in avg_spending_by_month_df.index.tolist()]
+
+        # get a dictionary of just the discretionary columns and how much was spent
+        disc_dict = dict(
+            avg_spending_by_month_df[avg_spending_by_month_df['disc'] == True]['mean'])
+
+        # rerverse dictionary so key is amount spent and value is category
+        disc_dict_reversed = {}
+        for k, v in disc_dict.items():
+            disc_dict_reversed[v] = k
+
+        # find the key:value pair that shows which discretionary category the user
+        # . spent the most money in
+        max_cat = max(disc_dict_reversed.items())
+
+        # subtract the monthly savings goal from that category
+        avg_cat_spending_dict[max_cat[1]] -= monthly_savings_goal
+
+        self.past_months = num_months
+
+        return avg_cat_spending_dict
+    
+    def current_month_spending(self):
+
+        cur_year = self.expenses['date'].max().year
+        cur_month = self.expenses['date'].max().month
+        user_exp = self.expenses.copy()
+        cur_month_expenses = user_exp[(user_exp['date'].dt.month == cur_month) &
+                                      (user_exp['date'].dt.year == cur_year)]
+        grouped_expenses = cur_month_expenses.groupby(
+            ['grandparent_category_name']).sum()
+        grouped_expenses = grouped_expenses.round({'amount_dollars': 2})
+
+        grouped_dict = dict(grouped_expenses['amount_dollars'])
+
+        # get dataframe of average spending per category over last 6 months
+        avg_spending_by_month_df = monthly_avg_spending(
+            self.expenses, num_months=self.past_months)
+
+        for cat in avg_spending_by_month_df.index:
+            if cat not in grouped_dict:
+                grouped_dict[cat] = 0
+
+        return grouped_dict
