+++ project/requirements.txt
@@ -5,3 +5,4 @@ uvicorn==0.11.8
 psycopg2-binary==2.8.5
 SQLAlchemy==1.3.19
 python-dotenv==0.14.0
+numpy==1.19.1
